#255 [QA] 헬스케어 오늘의 걸음수 업로드 시 소수점 절삭 되어 보내짐

+++ core/data/src/main/java/com/startup/data/remote/dto/request/healthcare/PutTodayWalkRequest.kt
@@ -8,7 +8,7 @@ data class PutTodayWalkRequest(
     @SerializedName("nowDay")
     val nowDay: String,
     @SerializedName("nowDistance")
-    val nowDistance: Int,
+    val nowDistance: Double,
     @SerializedName("nowSteps")
     val nowSteps: Int,
     @SerializedName("targetSteps")
